requirements.txt
Brotli==1.0.9
brotlipy==0.7.0
certifi==2021.5.30
colorama==0.4.4
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-core-components==1.17.1
dash-daq==0.5.0
dash-html-components==1.1.4
dash-renderer==1.9.0
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.5.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
networkx==2.6.2
openpyxl==3.0.7
pandas==1.3.1
plotly==5.1.0
PyYAML==5.4.1
retrying==1.3.3
scipy==1.7.1
tsp-solver2==0.4.1
wincertstore==0.2
